backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
openai
gunicorn
